+++ tests/testthat/test-find_python_path.R
@@ -1,15 +1,16 @@
 test_that("find_python_path works", {
-    conda_env <- env_from_yaml(yaml_path = system.file(
+    
+    conda_env <- echoconda::yaml_to_env(yaml_path = system.file(
         package = "echoconda",
         "conda/echoR.yml"
     ))
-
+    #### In Windows, the executable is named "python.exe" ####
     python <- find_python_path(conda_env = "echoR")
-    testthat::expect_equal(basename(python), "python")
+    testthat::expect_true(startsWith(basename(python),"python"))
 
     python <- find_python_path(conda_env = NULL)
-    testthat::expect_equal(basename(python), "python")
+    testthat::expect_true(startsWith(basename(python),"python"))
 
     python <- find_python_path(conda_env = "typo")
-    testthat::expect_equal(basename(python), "python")
+    testthat::expect_true(startsWith(basename(python),"python"))
 })