+++ NEWS.md
@@ -1,7 +1,16 @@
 # echoconda 0.99.4
 
-* Update GHA
+## New features
+
+* Update GitHub Actions. 
+* `find_packages`: New function to search for packages across all conda envs.
+* `env_to_yaml`: New function to export yaml file.
+* `yaml_to_env`: New function to create env from yaml file. 
+
+## Bug fixes 
+
 * Try to allow switching between envs again. 
+* Omit setting env vars in *zzz.R*. 
 
 # echoconda 0.99.0